+++ examples/cpp/qb_mpdo_noisy/qb_mpdo_noisy.cpp
@@ -25,6 +25,9 @@ int main()
   s.rel_svd_cutoff = 1.0e-10;
   s.measure_sample_method = "auto";
   s.gpu_device_ids = {0};
+  s.svd_type = "Jacobian";
+  s.svdj_tol = 1.0e-3;
+  s.svdj_max_sweeps = 80;
 
   // Set the noise model
   // Uncomment the following lines to introduce noise to the simulation.